### 2024-10-01 #662 @kelliejarcher
Guidance on whether hdcuremodels can be reviewed by ROpenSci 
Submitting Author Name: Kellie J. Archer
Submitting Author Github Handle: @kelliejarcher
Other Package Authors Github handles:
Repository:  https://github.com/kelliejarcher/hdcuremodels
Submission type: Pre-submission

---

```
Penalized Mixture Cure Models for High-Dimensional Data
```

When performing a time-to-event analysis to data that includes a subset of subjects who will not experience the event of interest (for convenience, cured), mixture cure models (MCMs) are a useful alternative to the Cox proportional hazards (PH) model because the Cox model assumes a constant hazard applies to all subjects throughout the observed follow-up time which is violated when some subjects in the dataset are cured and when cured subjects comprise a portion of the dataset, the survival function is improper. The `hdcuremodels` R package was developed for fitting penalized mixture cure models when there is a high-dimensional covariate space, such as when high-throughput genomic data are used in modeling time-to-event data when some subjects will not experience the event of interest.

I have gone through the list and identified items that have been addr
[hdcuremodels_0.0.2.tar.gz](https://github.com/user-attachments/files/17216601/hdcuremodels_0.0.2.tar.gz)
essed. I am unclear on several standards and others I am unclear how to implement.

Anyone who performs time-to-event analyses, particularly those who may have high-dimensional covariate spaces.

While there are a few R packages that can be used to fit mixture cure models, only `penPHcure`(Beretta and Heuchenne, 2019)  includes a LASSO penalty to perform variable selection for scenarios when the sample size exceeds the number of predictors and was introduced to fit semi-parametric proportional hazards mixture cure models with time-varying covariates. The hdcuremodels package can be used to fit penalized Weibull, exponential, and semi-parametric mixture cure models. Also, the hdcuremodels package includes functions for evaluating mixture cure model assumptions, namely to test whether there is a significant cure fraction and to test whether there is sufficient follow-up.

Yes

I did submit a version to CRAN back in June and subsequently learned about ROpenSci. My github site includes more recent updates where I have changed exported functions to use snake case, make use of match.args(), tolower, etc. I would like to improve this R package prior to submitting a paper to a peer-reviewed journal. Being new to ROpenSci I am unfamiliar with collaborating via github.
[hdcuremodels_0.0.2.tar.gz](https://github.com/user-attachments/files/17216609/hdcuremodels_0.0.2.tar.gz)
