### 2024-10-03 #662 @mpadge
Hi @kelliejarcher, I'm jumping in briefly while current Editor-in-Chief @adamhsparks is away on a short break. Your package definitely seems like a good fit. As a stats submission, assessing fit is easy: You'll just need to document compliance with at least half of all standards (both General and Regression), and then call `@ropensci-review-bot check srr` here to confirm.

In you're unsure about compliance with any particular standards, this pre-submission issue is the place to ask, and we'll do our best to guide you through until you've passed `check srr`. You can and should check locally as you document compliance, via [`srr::srr_stats_pre_submit()`](https://docs.ropensci.org/srr/reference/srr_stats_pre_submit.html), or [`srr::srr_report()`](https://docs.ropensci.org/srr/reference/srr_report.html) to generate full report.

To help you understand how you might best comply with standards, it might help to look through previous submissions in the *Regression and Supervised Learning* category, including #554, #550, and #603. I'd recommend that you clone any of those that look helpful, and run `srr::srr_report()` in the directory of that repo to generate full report for you to examine. Feel free to ask any more questions here.
