### 2024-10-31 #662 @adamhsparks
@kelliejarcher, my time as the EIC is ending today so I'm checking open issues to see if there are any I need to action to close. We're happy to keep this open if you have further discussion or questions for us about the process. Otherwise, are you happy for us to close this issue?
