### 2024-10-31 #662 @kelliejarcher
Thanks Adam. I have a few additional edits to make and will then be submitting my package for review. I am unclear whether this ticket needs to stay open for that submission or if that is a new submission. Either way is fine with me.

Best regards,
Kellie

From: Adam H. Sparks ***@***.***>
Date: Thursday, October 31, 2024 at 6:36 AM
To: ropensci/software-review ***@***.***>
Cc: Archer, Kellie ***@***.***>, Mention ***@***.***>
Subject: Re: [ropensci/software-review] Guidance on whether hdcuremodels can be reviewed by ROpenSci (Issue #662)
@ kelliejarcher, my time as the EIC is ending today so I'm checking open issues to see if there are any I need to action to close. We're happy to keep this open if you have further discussion or questions for us about the process. Otherwise,

@kelliejarcher<https://urldefense.com/v3/__https:/github.com/kelliejarcher__;!!KGKeukY!x7FxnoWc_c35XavfkusZ9xAc33cpKT29RFmEjwqlhXYipsvyr9UljUkhZI2K9eNHihs1t7Kct_6lx7QRtoTk3vp0Ng$>, my time as the EIC is ending today so I'm checking open issues to see if there are any I need to action to close. We're happy to keep this open if you have further discussion or questions for us about the process. Otherwise, are you happy for us to close this issue?
