### 2024-11-02 #662 @adamhsparks
@kelliejarcher, the submission form for a review request asks if you've submitted a pre-submission request and for you to provide a link to it and/or tag the handling editor of that request.

cheers!
